New Development: No             JIRA Issue: Yes CAS-6446         Ready for Test: Yes      Interface Changes: Yes/No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes/No              Module(s): Module Names change impacts.            Description: Describe your changes here...
DetectLine is updated so that it takes masks as an input parameter 
and handles it properly. Masks are passed from MaskLine and are 
based on spectral data from SimpleGriddingResults. Masks are 
determined by the following calculation, 

    masks = (spectra != NoData)

where spectra is a two-dimensional (nrow,nchan) spectral data 
taken from SimpleGriddingResults, and NoData is a special value 
that is used when no data is accumulated during gridding. The 
data is valid if mask is True. 


	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/project/pipeline@29818 ac9c0da5-131f-0410-9009-a9c195f183e6