Removed calmaxpix parameter from hif_editimlist
'calmaxpix' is used in hif_makeimlist to override the
automatic imsize heuristics for calibrators, but is not
needed in hif_editimlist because imsize is set manually.

I also removed unused imports, commented out code,
and the _DESCRIPTIONS dictionary for the weblog until
we decide if it is needed.

Finally, I cleaned up some of the XML help text and
did a little code beautification.



git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/project/pipeline@39467 ac9c0da5-131f-0410-9009-a9c195f183e6