CAS-12076: introduces a new parameter exportcalprods to hifv_exportdata.
The new parameter, exportcalprods, is a subparameter of exportmses and valid when exportmses = True.
When both exportmses and exportcalprods are True, both calibration products (e.g., caltables, restorescripts, flagversion) and MSes are exported to product directory.
The default behavioro of hifv_exportdata is unchanged. By default, exportmses = False and in this case, only calibration products are exported.

git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/project/pipeline@42287 ac9c0da5-131f-0410-9009-a9c195f183e6