PIPE-334 *IMPORTANT BUG FIX* on maximum line width for detected lines Previously, lines were rejected even if it was detected as a line either
(1) when line width exceeded 900ch or 
(2) when line width exceeded 33% of effective number of channels

However, the condition (1) may do harm when nchan becomes large (e.g. 4096ch) 
so that such kind of hard-coded threshold must be avoided. For (2), it was 
found that 33% is too tight condition for rejection. 

As a result, we decided to delete the condition (1) and relax the condition (2) 
from 33% to 50%. Detailed report will be provided by George Kosugi later.

git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/project/Pipeline-CASA56-P1-B@42823 ac9c0da5-131f-0410-9009-a9c195f183e6