PIPE-10: updates for VLASS imaging
The stringVec type for the imsize parameter was not behaving nicely with casa6.
I changed the type to intArray, which is probably a more apt type for this
parameter anyway.

I also created a utility function to handle None values with quanta.quantity()
These were previsously acceptable in casa5 but are not allowed in casa6.