PIPE-618 hif_makeimlist weblog should not show nterms column
nterms column cannot be filled correctly in hif_makeimlist, as the decision is made
later in the hif_makeimages task. The column was shown always in the weblog, because
CleanTarget object always has nterms argument. The statement for filling the column
however was missing from makeimlist.mako template. With this commit the nterms column
is altogether removed from the template.