PIPE-1681 fix unexpected edit of casa_pipescript.py
The code to remove output_dir from casa_pipescript.py edits the file unexpectedly
when output_dir is empty string. This commit intends to prevent it.

Use of os.path.isdir is inspired by Executor._log_jobrequest. This way, output_dir
is removed only when output_dir is non-empty and it points to existing directory.
This is more robust than simple string check like "if output_dir" or "if len(output_dir) > 0".i