PIPE-2009: create a new row div once the row is full
Plot wrapping did not work with the unusual dimensions of the hifa_tsysflag continuum diagnostic plots, with
the symptom being the second row did not have the plots flushed left. This patch inserts a new div class="row"
HTML element every four plots, which restores the expected alignment. This logic requires the number of plots
to be fixed to four per row; as a result, the switching to 3 plots per row on smaller screens (<768px) is
disabled - although small screen display never functioned correctly with the navbar so this is not considered
a loss of functionality.