Fix scorecalculator unit test failure
It seems that builtin sum in Python 3.10 tries to keep data type of the values to be summed up even if it causes overflow. On the other hand, np.sum flexibly changes data type of resulting value when data type is integer and the resulting value causes overflow with the input data type.