PIPE-1770: fix QA score overwrite bug in hifa_applycal summarisation
Summarisation of hifa_applycal QA scores would incorrectly handle scores
with the same data selection. This patch fixes the issue so that all
scores with a given data selection are considered by adding them all to
a list, rather than the previous behaviour where the list was repeatedly
overwritten with list containing the score in scope.