PIPE-2791: removed unnecessary 'if' statements in ms_preable.
This commit handles Item 3 in the description of PIPE-2791.
Removed a couple of "if uv_max[ms].value > 0.0" statements in ms_preables
in applycal.mako. The 'if' statements define texts (description of plots)
shown only for interferometry runs.
However those plots are not generated in single dish runs. Hence, the change
does not impact appearlance of weblog.