PIPE-2791: removed unused codes in hsdn_restoredata.mako.
This commit handles Item 4 in the description of PIPE-2791.
Mako template of hsdn_restoredata task has codes for plots that are not
created in hsdn_restoredata task. It is because the template was copied
from that of applycal. The commit removes the codes that are not used
in hsdn_restoredata.